runtime/samples/nsnet2/requirements.txt
torch==2.3.0
iree-turbine==2.3.0
